refactor(backend): move into separate module
